SAAS/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pypdf
python-docx
pytest
httpx

